+++ integration_tests/tests/end_to_end.rs
@@ -1,76 +1,26 @@
-use std::panic::AssertUnwindSafe;
-use std::sync::LazyLock;
-use std::time::Instant;
-use std::{
-    collections::HashMap,
-    sync::{Arc, RwLock},
-};
-
 use bb8_postgres::PostgresConnectionManager;
 use chrono::{DateTime, Duration, DurationRound, TimeDelta, TimeZone, Utc};
 use chronoutil::RelativeDuration;
-use futures::{Future, FutureExt};
-use rdkafka::producer::{FutureProducer, FutureRecord};
 use rove::data_switch::{DataConnector, SpaceSpec, TimeSpec, Timestamp};
-use tokio::task::JoinHandle;
 use tokio_postgres::NoTls;
-use tokio_util::sync::CancellationToken;
 
 use lard_egress::{timeseries::Timeseries, LatestResp, TimeseriesResp, TimesliceResp};
-use lard_ingestion::{
-    levels::{param_get_level, Level, ParamLevelTable},
-    permissions::{timeseries_get_permit, ParamPermit, ParamPermitTable, StationPermitTable},
-    qc_pipelines::load_pipelines,
-    DbPools, KldataResp,
-};
-use rove_connector::Connector;
+use lard_ingestion::{levels::param_get_level, permissions::timeseries_get_permit, KldataResp};
 
-const CONNECT_STRING_LARD: &str = "host=localhost user=postgres dbname=lard password=postgres";
-const CONNECT_STRING_LARD_RESTRICTED: &str =
-    "host=localhost user=postgres dbname=lard_restricted password=postgres";
-const PARAMCONV_CSV: &str = "../resources/paramconversions.csv";
-const KAFKA_TOPIC: &str = "checked";
-const KAFKA_GROUP: &str = "lard_test";
-
-// TODO: make API and ingestor global static as well? So we don't have to recreate them for each test?
-static PARAMETERS: LazyLock<HashMap<String, (i32, TestObsType)>> = LazyLock::new(|| {
-    csv::Reader::from_path(PARAMCONV_CSV)
-        .unwrap()
-        .into_records()
-        .map(|record_result| {
-            let record = record_result.unwrap();
-            (
-                record.get(1).unwrap().to_owned(),
-                (
-                    record.get(0).unwrap().parse::<i32>().unwrap(),
-                    match record.get(3).unwrap() {
-                        "t" => TestObsType::Scalar,
-                        "f" => TestObsType::NonScalar,
-                        _ => unreachable!(),
-                    },
-                ),
-            )
-        })
-        .collect()
-});
-
-#[derive(Clone, Copy)]
-enum TestObsType {
-    Scalar,
-    NonScalar,
-}
+pub mod common;
+use common::e2e_test_wrapper;
 
 #[derive(Clone)]
 struct Param<'a> {
     id: i32,
     code: &'a str,
     sensor_level: Option<(i32, i32)>,
-    obstype: TestObsType,
+    obstype: common::TestObsType,
 }
 
 impl Param<'_> {
     fn new(code: &str) -> Self {
-        let (code, (id, obstype)) = PARAMETERS
+        let (code, (id, obstype)) = common::PARAMETERS
             .get_key_value(code)
             .expect("Provided param code should be present in global params hashmap");
 
@@ -83,7 +33,7 @@ impl Param<'_> {
     }
 
     fn with_sensor_level(code: &str, sensor_level: (i32, i32)) -> Self {
-        let (code, (id, obstype)) = PARAMETERS
+        let (code, (id, obstype)) = common::PARAMETERS
             .get_key_value(code)
             .expect("Provided param code should be present in global params hashmap");
 
@@ -122,8 +72,8 @@ impl TestData<'_> {
             .params
             .iter()
             .map(|param| match param.obstype {
-                TestObsType::Scalar => scalar_val.to_string(),
-                TestObsType::NonScalar => nonscalar_val.to_string(),
+                common::TestObsType::Scalar => scalar_val.to_string(),
+                common::TestObsType::NonScalar => nonscalar_val.to_string(),
             })
             .collect::<Vec<String>>()
             .join(",");
@@ -163,61 +113,6 @@ impl TestData<'_> {
     }
 }
 
-fn mock_level_table() -> Arc<RwLock<ParamLevelTable>> {
-    let param_level = HashMap::from([
-        (211, Level::new(2, 0)),
-        (81, Level::new(10, 0)),
-        (3, Level::new(20, -2)),
-    ]);
-
-    Arc::new(RwLock::new(param_level))
-}
-
-#[test]
-fn test_param_get_level() {
-    let cases = vec![
-        (211, Some(0), 200, "air_temperature default is 2m"),
-        (
-            211,
-            Some(10),
-            1000,
-            "air_temperature at 10m converted to cm",
-        ),
-        (81, Some(0), 1000, "wind_speed default is 10m"),
-        (3, Some(0), 20, "3 default is 20cm"),
-    ];
-
-    let level_table = mock_level_table();
-    for case in cases {
-        let param_id = case.0;
-        let level = case.1;
-        let expected = case.2;
-        let test_case = case.3;
-
-        let output = param_get_level(level_table.clone(), param_id, level).unwrap();
-        assert_eq!(output, Some(expected), "{}", test_case);
-    }
-}
-
-fn mock_permit_tables() -> Arc<RwLock<(ParamPermitTable, StationPermitTable)>> {
-    let param_permit = HashMap::from([
-        // station_id -> (type_id, param_id, permit_id)
-        (10000, vec![ParamPermit::new(0, 0, 0)]),
-        (10001, vec![ParamPermit::new(0, 0, 1)]), // open
-    ]);
-
-    let station_permit = HashMap::from([
-        // station_id -> permit_id
-        (10000, 1), // overridden by param_permit
-        (10001, 0), // overridden by param_permit
-        (20000, 0),
-        (20001, 1), // open
-        (20002, 1), // open
-    ]);
-
-    Arc::new(RwLock::new((param_permit, station_permit)))
-}
-
 #[test]
 fn test_timeseries_get_permit() {
     let cases = vec![
@@ -253,7 +148,7 @@ fn test_timeseries_get_permit() {
         ),
     ];
 
-    let permit_tables = mock_permit_tables();
+    let permit_tables = common::mock_permit_tables();
     for case in cases {
         let station_id = case.0;
         let type_id = case.1;
@@ -267,129 +162,30 @@ fn test_timeseries_get_permit() {
     }
 }
 
-async fn wrapper_setup() -> (DbPools, JoinHandle<()>, CancellationToken) {
-    let open_manager =
-        PostgresConnectionManager::new_from_stringlike(CONNECT_STRING_LARD, NoTls).unwrap();
-    let open_db_pool = bb8::Pool::builder().build(open_manager).await.unwrap();
-    let restricted_manager =
-        PostgresConnectionManager::new_from_stringlike(CONNECT_STRING_LARD_RESTRICTED, NoTls)
-            .unwrap();
-    let restricted_db_pool = bb8::Pool::builder()
-        .build(restricted_manager)
-        .await
-        .unwrap();
+#[test]
+fn test_param_get_level() {
+    let cases = vec![
+        (211, Some(0), 200, "air_temperature default is 2m"),
+        (
+            211,
+            Some(10),
+            1000,
+            "air_temperature at 10m converted to cm",
+        ),
+        (81, Some(0), 1000, "wind_speed default is 10m"),
+        (3, Some(0), 20, "3 default is 20cm"),
+    ];
 
-    let egress_pool = open_db_pool.clone();
-    let db_pools = DbPools {
-        open: open_db_pool.clone(),
-        restricted: restricted_db_pool.clone(),
-    };
+    let level_table = common::mock_level_table();
+    for case in cases {
+        let param_id = case.0;
+        let level = case.1;
+        let expected = case.2;
+        let test_case = case.3;
 
-    // set up cancellation token and signal catcher to detect premature shutdown
-    let cancel_token = CancellationToken::new();
-
-    let egress = tokio::spawn(lard_egress::run(egress_pool, cancel_token.clone()));
-
-    (db_pools, egress, cancel_token)
-}
-
-async fn db_cleanup(db_pools: DbPools) {
-    for db_pool in [db_pools.open, db_pools.restricted] {
-        let client = db_pool.get().await.unwrap();
-        client
-            .batch_execute(
-                // TODO: should clean public.timeseries_id_seq too? RESTART IDENTITY CASCADE?
-                "TRUNCATE public.timeseries, labels.met, labels.obsinn CASCADE",
-            )
-            .await
-            .unwrap();
+        let output = param_get_level(level_table.clone(), param_id, level).unwrap();
+        assert_eq!(output, Some(expected), "{}", test_case);
     }
-}
-
-async fn e2e_test_wrapper_kldata<T: Future<Output = ()>>(test: T) {
-    let (db_pools, mut egress, cancel_token) = wrapper_setup().await;
-
-    let rove_connector = Connector {
-        pool: db_pools.open.clone(),
-    };
-    let qc_pipelines = load_pipelines("mock_qc_pipelines/fresh").expect("failed to load pipelines");
-
-    let mut ingestion = tokio::spawn(lard_ingestion::run(
-        db_pools.clone(),
-        PARAMCONV_CSV,
-        mock_permit_tables(),
-        mock_level_table(),
-        rove_connector,
-        qc_pipelines,
-        cancel_token.clone(),
-    ));
-
-    tokio::select! {
-        _ = &mut egress => panic!("API server task terminated first"),
-        _ = &mut ingestion => panic!("Ingestor server task terminated first"),
-        // Clean up database even if test panics, to avoid test poisoning
-        test_result = AssertUnwindSafe(test).catch_unwind() => {
-            // For debugging a specific test, it might be useful to skip the cleanup process
-            #[cfg(not(feature = "debug"))]
-            db_cleanup(db_pools).await;
-
-            assert!(test_result.is_ok())
-        }
-    }
-
-    cancel_token.cancel();
-    let (egress_result, ingestion_result) = tokio::join!(egress, ingestion);
-    egress_result.unwrap();
-    ingestion_result.unwrap().unwrap()
-}
-
-/// Similar to e2e_test_wrapper, but adapted to use kvkafka ingestion instead of obsinn.
-async fn e2e_test_wrapper_kvkafka(test: impl AsyncFnOnce(FutureProducer, DbPools) -> ()) {
-    let (db_pools, mut egress, cancel_token) = wrapper_setup().await;
-
-    let mock_kafka_cluster = rdkafka::mocking::MockCluster::new(3).unwrap();
-    mock_kafka_cluster.create_topic(KAFKA_TOPIC, 32, 3).unwrap();
-    let kafka_brokers = mock_kafka_cluster.bootstrap_servers();
-
-    let kafka_producer: FutureProducer = rdkafka::ClientConfig::new()
-        .set("bootstrap.servers", kafka_brokers.clone())
-        .create()
-        .unwrap();
-
-    let (ingestion_pools, ingestion_token) = (db_pools.clone(), cancel_token.clone());
-    let mut ingestion = tokio::spawn(async move {
-        let kafka_brokers = kafka_brokers;
-        lard_ingestion::kvkafka::ingest_kvkafka(
-            ingestion_pools,
-            &kafka_brokers,
-            KAFKA_GROUP,
-            KAFKA_TOPIC,
-            ingestion_token,
-            mock_permit_tables(),
-            mock_level_table(),
-        )
-        .await
-    });
-
-    tokio::select! {
-        _ = &mut egress => panic!("API server task terminated first"),
-        _ = &mut ingestion => panic!("Ingestor server task terminated first"),
-        // Clean up database even if test panics, to avoid test poisoning
-        test_result = AssertUnwindSafe(test(kafka_producer, db_pools.clone())).catch_unwind() => {
-            // For debugging a specific test, it might be useful to skip the cleanup process
-            #[cfg(not(feature = "debug"))]
-            if test_result.is_err() {
-                db_cleanup(db_pools.clone()).await;
-            }
-
-            assert!(test_result.is_ok())
-        }
-    }
-
-    cancel_token.cancel();
-    let (egress_result, ingestion_result) = tokio::join!(egress, ingestion);
-    egress_result.unwrap();
-    ingestion_result.unwrap().unwrap();
 }
 
 async fn ingest_data(client: &reqwest::Client, obsinn_msg: String) -> KldataResp {
@@ -405,7 +201,7 @@ async fn ingest_data(client: &reqwest::Client, obsinn_msg: String) -> KldataResp
 
 #[tokio::test]
 async fn test_stations_endpoint_irregular() {
-    e2e_test_wrapper_kldata(async {
+    e2e_test_wrapper(async {
         let ts = TestData {
             station_id: 20001,
             params: vec![Param::new("TGM"), Param::new("TGX")],
@@ -477,7 +273,7 @@ async fn test_stations_endpoint_regular() {
     ];
 
     for ts in cases {
-        e2e_test_wrapper_kldata(async {
+        e2e_test_wrapper(async {
             let client = reqwest::Client::new();
             let ingestor_resp = ingest_data(&client, ts.obsinn_message()).await;
             assert_eq!(ingestor_resp.res, 0);
@@ -513,7 +309,7 @@ async fn test_stations_endpoint_errors() {
         (20001, 999),
     ];
     for (station_id, param_id) in cases {
-        e2e_test_wrapper_kldata(async {
+        e2e_test_wrapper(async {
             let ts = TestData {
                 station_id: 20001,
                 params: vec![Param::new("TA")],
@@ -553,7 +349,7 @@ async fn test_latest_endpoint() {
         ("?latest_max_age=2019-01-01T00:00:00Z", 4),
     ];
     for (query, n_timeseries_found) in cases {
-        e2e_test_wrapper_kldata(async {
+        e2e_test_wrapper(async {
             let test_data = [
                 TestData {
                     station_id: 20001,
@@ -593,7 +389,7 @@ async fn test_latest_endpoint() {
 
 #[tokio::test]
 async fn test_timeslice_endpoint() {
-    e2e_test_wrapper_kldata(async {
+    e2e_test_wrapper(async {
         let timestamp = Utc.with_ymd_and_hms(2024, 1, 1, 1, 0, 0).unwrap();
         let params = vec![Param::new("TA")];
 
@@ -652,156 +448,6 @@ async fn test_timeslice_endpoint() {
 }
 
 #[tokio::test]
-async fn test_kafka() {
-    e2e_test_wrapper_kvkafka(async |producer: FutureProducer, db_pools: DbPools| {
-        // This observation was 2.5 hours late??
-        let kafka_xml = r#"<?xml?>
-            <KvalobsData producer=\"kvqabase\" created=\"2024-06-06 08:30:43\">
-                <station val=\"20001\">
-                    <typeid val=\"-4\">
-                        <obstime val=\"2024-06-06 06:00:00\">
-                            <tbtime val=\"2024-06-06 08:30:42.943247\">
-                                <sensor val=\"0\">
-                                    <level val=\"0\">
-                                        <kvdata paramid=\"106\">
-                                            <original>10</original>
-                                            <corrected>10</corrected>
-                                            <controlinfo>1000000000000000</controlinfo>
-                                            <useinfo>9000000000000000</useinfo>
-                                            <cfailed></cfailed>
-                                        </kvdata>
-                                    </level>
-                                </sensor>
-                            </tbtime>
-                        </obstime>
-                    </typeid>
-                </station>
-            </KvalobsData>"#;
-
-        producer
-            .send_result(FutureRecord::to(KAFKA_TOPIC).key("").payload(kafka_xml))
-            .unwrap()
-            .await
-            .unwrap()
-            .unwrap();
-
-        // TODO: we do not have an API endpoint to query the flags.kvdata table
-        let open_conn = db_pools.open.get().await.unwrap();
-
-        // As we have no way to sync with message processing in kvkafka ingestion, we just keep
-        // trying to fetch data with a timeout
-        let timeout = std::time::Duration::from_secs(10);
-        let timeout_start = Instant::now();
-        loop {
-            if let Ok(data_row) = open_conn
-                .query_one(
-                    r#"
-                        SELECT
-                            timeseries,
-                            obstime,
-                            original,
-                            corrected,
-                            quality_code,
-                            controlinfo,
-                            useinfo,
-                            cfailed
-                        FROM legacy.data
-                    "#,
-                    &[],
-                )
-                .await
-            {
-                #[allow(clippy::type_complexity)]
-                let (
-                    timeseries,
-                    obstime,
-                    original,
-                    corrected,
-                    quality_code,
-                    controlinfo,
-                    useinfo,
-                    cfailed,
-                ): (
-                    i64,
-                    DateTime<Utc>,
-                    Option<f64>,
-                    Option<f64>,
-                    Option<i32>,
-                    Option<String>,
-                    Option<String>,
-                    Option<String>,
-                ) = (
-                    data_row.get(0),
-                    data_row.get(1),
-                    data_row.get(2),
-                    data_row.get(3),
-                    data_row.get(4),
-                    data_row.get(5),
-                    data_row.get(6),
-                    data_row.get(7),
-                );
-                assert_eq!(obstime, Utc.with_ymd_and_hms(2024, 6, 6, 6, 0, 0).unwrap());
-                assert_eq!(original, Some(10.));
-                assert_eq!(corrected, Some(10.));
-                assert_eq!(
-                    quality_code,
-                    lard_ingestion::kvkafka::get_quality_code(useinfo.clone().unwrap().as_str())
-                );
-                assert_eq!(controlinfo, Some("1000000000000000".to_string()));
-                assert_eq!(useinfo, Some("9000000000000000".to_string()));
-                assert_eq!(cfailed, None);
-
-                let label_row = open_conn
-                    .query_one(
-                        r#"
-                        SELECT
-                            station_id,
-                            param_id,
-                            type_id,
-                            lvl,
-                            sensor
-                        FROM labels.kvalobs
-                        WHERE timeseries = $1
-                    "#,
-                        &[&timeseries],
-                    )
-                    .await
-                    .unwrap();
-
-                #[allow(clippy::type_complexity)]
-                let (station_id, param_id, type_id, lvl, sensor): (
-                    // should these really all be Option??
-                    Option<i32>,
-                    Option<i32>,
-                    Option<i32>,
-                    Option<i32>,
-                    Option<i32>,
-                ) = (
-                    label_row.get(0),
-                    label_row.get(1),
-                    label_row.get(2),
-                    label_row.get(3),
-                    label_row.get(4),
-                );
-
-                assert_eq!(station_id, Some(20001));
-                assert_eq!(param_id, Some(106));
-                assert_eq!(type_id, Some(-4));
-                assert_eq!(lvl, Some(0));
-                assert_eq!(sensor, Some(0));
-
-                break;
-            }
-
-            if timeout_start.elapsed() > timeout {
-                panic!("Timed out waiting for data to appear")
-            }
-        }
-    })
-    .await
-}
-
-#[tokio::test]
 async fn test_rove_connector() {
     let ts = TestData {
         station_id: 20001,
@@ -812,11 +458,12 @@ async fn test_rove_connector() {
         len: 12,
     };
 
-    e2e_test_wrapper_kldata(async {
+    e2e_test_wrapper(async {
         let client = reqwest::Client::new();
 
         let manager =
-            PostgresConnectionManager::new_from_stringlike(CONNECT_STRING_LARD, NoTls).unwrap();
+            PostgresConnectionManager::new_from_stringlike(common::CONNECT_STRING_LARD, NoTls)
+                .unwrap();
         let pool = bb8::Pool::builder().build(manager).await.unwrap();
         let connector = rove_connector::Connector { pool };
